dia_00/python_setup/requirements.txt
elasticsearch==8.11.1
python-dotenv
